[WIP][RISCV] Insert vsetvli before whole register moves
This is very much a proof of concept patch.  It is incorrect - as copies
inserted or moved after vsetvli insertion are not handled.  It is
overly conservative - as I abuse existing tracking, and we appear to
be missing a bunch of cases where we can rewrite existing vsetvlis, and
don't kill vsetvlis inserted when copies are removed.

I wrote this mostly to get a feel for the impact on codegen a couple
months back.  Sharing it because it was requested.  I do not plan to
drive this towards submission.