[RISCV] Handle length reducing shuffles in costShuffleViaVRegSplitting
The previous code checked to see if the number of destinations was 1,
but didn't check if the number of sources was also 1.  If we have
more sources than destinations, we have to widen to wider of the two
when performing an arbitrary shuffle.

Note that not all shuffles are arbitrary - a deinterleave which could
be lowered to a vnsrl might also get picked up by this.  This is
undesirable from the minimum costing perspective, but does match what
our current lowering will currently do.  We don't see this in test
deltas because getInstructionCost doesn't handle length changing
shuffles well, but this difference might be visible through e.g.
direct getShuffleCost queries from SLP.